cif report

+++ src/main/java/com/cimb/vipflag/dao/CifDataUploadedRepo.java
@@ -1,13 +1,11 @@
 package com.cimb.vipflag.dao;
 
-import com.cimb.vipflag.entity.CifDataChecksum;
 import com.cimb.vipflag.entity.CifDataUploaded;
-import com.cimb.vipflag.entity.FileDirectory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 public interface CifDataUploadedRepo extends JpaRepository<CifDataUploaded, Integer> {
 
     @Query(value = "select * from cif_data_uploaded where cfcifn=?1",nativeQuery = true)
-    public CifDataUploaded findCFCIFN(int cfcifn );
+    public CifDataUploaded findCFCIFN(long cfcifn );
 }
